NEXUS-4904: Requiring PathCache in NexusConfiguration component.
This is a workaround that matters only during UTs, as almost all of them
from nexus-app module "onwards" will refer (directly or indirectly) to this
component, and by having this requirement, indirectly to PathCache.
Again, on runtime/deployment, this change does not matter as PathCache
is already in component dependency graph, but in UTs this simple
change enforces Plexus to apply lifecycle, and hence, manage EhCache
manager, as since 2.5 it enforces strictly that a manager has to be
shut down before another one (with same name, something we never manager)
is started up.